requirements.txt
openai
tiktoken
numpy
pandas
python-dotenv
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary